Refactor: architecture
