+++ back-end/src/services/productsServices.js
@@ -2,7 +2,7 @@ const { Product } = require('../database/models');
 
 const getAllProductsService = async () => {
   const products = await Product.findAll();
-  console.log(Product, 'ba', products);
+
   return products;
 };
