+++ front-end/src/common/components/CheckoutComponents/CheckoutButtonProducts.jsx
@@ -1,9 +1,9 @@
 import React, { useContext } from 'react';
-import '../styles/CheckoutButtonProducts.css';
+import '../../../styles/CheckoutButtonProducts.css';
 import { Link } from 'react-router-dom';
-import { Context } from '../provider/Provider';
+import { Context } from '../../../provider/Provider';
 
-import addZeroes from '../helper/functions/addZeroes';
+import addZeroes from '../../../helper/functions/addZeroes';
 
 const CheckoutButtonProducts = () => {
   const { totalOrder } = useContext(Context);
