+++ front-end/src/common/components/CheckoutComponents/CheckoutContainer.jsx
@@ -4,8 +4,8 @@ import CheckoutProducts from './CheckoutProducts';
 import PurshaseTotal from './PurchaseTotal';
 import PurchaseOrderBtn from './PurchaseOrderBtn';
 import DeliveryContainer from './DeliveryContainer';
-import { Context } from '../provider/Provider';
-import Header from './Header';
+import { Context } from '../../../provider/Provider';
+import Header from '../Header/Header';
 
 function CheckoutContainer() {
   const { orderInProgress } = React.useContext(Context);
