+++ front-end/src/common/components/CheckoutComponents/EntryAddress.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Context } from '../provider/Provider';
+import { Context } from '../../../provider/Provider';
 
 function EntryAddress() {
   const { setPurchaseAddress } = React.useContext(Context);
