+++ front-end/src/common/components/CheckoutComponents/PurchaseOrderBtn.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Context } from '../provider/Provider';
+import { Context } from '../../../provider/Provider';
 
 function PurchaseOrderBtn() {
   const { checkoutPurchase } = React.useContext(Context);
