+++ front-end/src/common/components/CheckoutComponents/PurchaseTotal.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { Context } from '../provider/Provider';
+import { Context } from '../../../provider/Provider';
 
 function PurshaseTotal() {
   const { orderInProgress, totalOrder, setTotalOrder } = React.useContext(Context);
