+++ front-end/src/common/components/ProductsComponents/ProductsCards.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
-import { Context } from '../provider/Provider';
-import fixPrice from '../helper/functions/fixPrice';
-import '../styles/ProductsCards.css';
+import { Context } from '../../../provider/Provider';
+import fixPrice from '../../../helper/functions/fixPrice';
+import '../../../styles/ProductsCards.css';
 import ProductsCardsHandleButtons from './ProductsCardsHandleButtons';
 
 const ProductsCards = () => {
