+++ front-end/src/common/pages/checkout/Checkout.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 /* import { Context } from '../../../provider/Provider'; */
-import CheckoutContainer from '../../../components/CheckoutContainer';
+import CheckoutContainer from '../../components/CheckoutComponents/CheckoutContainer';
 
 function Checkout() {
   /* const { totalOrder } = React.useContext(Context); */
