+++ front-end/src/common/pages/products/Products.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
-import Header from '../../../components/Header';
+import Header from '../../components/Header/Header';
 import '../../../styles/Products.css';
-import ProductsCards from '../../../components/ProductsCards';
-import CheckoutButtonProducts from '../../../components/CheckoutButtonProducts';
+import ProductsCards from '../../components/ProductsComponents/ProductsCards';
+import CheckoutButtonProducts from
+  '../../components/ProductsComponents/ProductsCardsHandleButtons';
 
 const Products = () => (
   <div className="main-div">
